requirements.txt
aiodns
aiohttp
aiosignal
async-timeout
attrs
Brotli
cffi
charset-normalizer
cryptg
faust-cchardet
frozenlist
hachoir
idna
multidict
Pillow
pyaes
pyasn1
pycares
pycparser
rsa
Telethon
yarl
PyYAML
dependency-injector